Move upgrade related code to its own class
