+++ sidekiq_queue_metrics.gemspec
@@ -19,6 +19,7 @@ Gem::Specification.new do |s|
 
   s.add_dependency                  'sidekiq', '>= 5.1'
   s.add_dependency                  'eldritch'
+  s.add_dependency                  'redlock'
 
   s.add_development_dependency      'bundler', '~> 1.5'
   s.add_development_dependency      'rspec'
